starting to make changes for book 2 examples for swift 3

+++ bk2ch01p006customWindowInStoryboardApp/ch14p366customWindowInStoryboardApp/ViewController.swift
@@ -22,15 +22,15 @@ class ViewController : UIViewController {
     override func viewDidLoad() {
         let mainview = self.view
         let lay1 = CALayer()
-        lay1.backgroundColor = UIColor(red: 1.0 as CGFloat, green: 0.4, blue: 1, alpha: 1).cgColor
+        lay1.backgroundColor = UIColor(red: 1.0, green: 0.4, blue: 1, alpha: 1).cgColor
         lay1.frame = CGRect(113, 111, 132, 194)
         mainview.layer.addSublayer(lay1)
         let lay2 = CALayer()
-        lay2.backgroundColor = UIColor(red: 0.5 as CGFloat, green: 1, blue: 0, alpha: 1).cgColor
+        lay2.backgroundColor = UIColor(red: 0.5, green: 1, blue: 0, alpha: 1).cgColor
         lay2.frame = CGRect(41, 56, 132, 194)
         lay1.addSublayer(lay2)
         let lay3 = CALayer()
-        lay3.backgroundColor = UIColor(red: 1 as CGFloat, green: 0, blue: 0, alpha: 1).cgColor
+        lay3.backgroundColor = UIColor(red: 1, green: 0, blue: 0, alpha: 1).cgColor
         lay3.frame = CGRect(43, 197, 160, 230)
         mainview.layer.addSublayer(lay3)
