+++ bk2ch01p017transform/ch14p371frame/AppDelegate.swift
@@ -25,9 +25,9 @@ extension CGRect {
         switch which {
         case 1:
             let v1 = UIView(frame:CGRect(113, 111, 132, 194))
-            v1.backgroundColor = UIColor(red: 1 as CGFloat, green: 0.4, blue: 1, alpha: 1)
+            v1.backgroundColor = UIColor(red: 1, green: 0.4, blue: 1, alpha: 1)
             let v2 = UIView(frame:v1.bounds.insetBy(dx: 10, dy: 10))
-            v2.backgroundColor = UIColor(red: 0.5 as CGFloat, green: 1, blue: 0, alpha: 1)
+            v2.backgroundColor = UIColor(red: 0.5, green: 1, blue: 0, alpha: 1)
             mainview.addSubview(v1)
             v1.addSubview(v2)
             
@@ -36,9 +36,9 @@ extension CGRect {
             
         case 2:
             let v1 = UIView(frame:CGRect(113, 111, 132, 194))
-            v1.backgroundColor = UIColor(red: 1 as CGFloat, green: 0.4, blue: 1, alpha: 1)
+            v1.backgroundColor = UIColor(red: 1, green: 0.4, blue: 1, alpha: 1)
             let v2 = UIView(frame:v1.bounds.insetBy(dx: 10, dy: 10))
-            v2.backgroundColor = UIColor(red: 0.5 as CGFloat, green: 1, blue: 0, alpha: 1)
+            v2.backgroundColor = UIColor(red: 0.5, green: 1, blue: 0, alpha: 1)
             mainview.addSubview(v1)
             v1.addSubview(v2)
             
@@ -46,9 +46,9 @@ extension CGRect {
             
         case 3:
             let v1 = UIView(frame:CGRect(20, 111, 132, 194))
-            v1.backgroundColor = UIColor(red: 1 as CGFloat, green: 0.4, blue: 1, alpha: 1)
+            v1.backgroundColor = UIColor(red: 1, green: 0.4, blue: 1, alpha: 1)
             let v2 = UIView(frame:v1.bounds)
-            v2.backgroundColor = UIColor(red: 0.5 as CGFloat, green: 1, blue: 0, alpha: 1)
+            v2.backgroundColor = UIColor(red: 0.5, green: 1, blue: 0, alpha: 1)
             mainview.addSubview(v1)
             v1.addSubview(v2)
             
@@ -57,9 +57,9 @@ extension CGRect {
             
         case 4:
             let v1 = UIView(frame:CGRect(20, 111, 132, 194))
-            v1.backgroundColor = UIColor(red: 1 as CGFloat, green: 0.4, blue: 1, alpha: 1)
+            v1.backgroundColor = UIColor(red: 1, green: 0.4, blue: 1, alpha: 1)
             let v2 = UIView(frame:v1.bounds)
-            v2.backgroundColor = UIColor(red: 0.5 as CGFloat, green: 1, blue: 0, alpha: 1)
+            v2.backgroundColor = UIColor(red: 0.5, green: 1, blue: 0, alpha: 1)
             mainview.addSubview(v1)
             v1.addSubview(v2)
             
@@ -68,9 +68,9 @@ extension CGRect {
             
         case 5: // same as case 4 but using concat
             let v1 = UIView(frame:CGRect(20, 111, 132, 194))
-            v1.backgroundColor = UIColor(red: 1 as CGFloat, green: 0.4, blue: 1, alpha: 1)
+            v1.backgroundColor = UIColor(red: 1, green: 0.4, blue: 1, alpha: 1)
             let v2 = UIView(frame:v1.bounds)
-            v2.backgroundColor = UIColor(red: 0.5 as CGFloat, green: 1, blue: 0, alpha: 1)
+            v2.backgroundColor = UIColor(red: 0.5, green: 1, blue: 0, alpha: 1)
             mainview.addSubview(v1)
             v1.addSubview(v2)
             
@@ -80,9 +80,9 @@ extension CGRect {
             
         case 6:
             let v1 = UIView(frame:CGRect(20, 111, 132, 194))
-            v1.backgroundColor = UIColor(red: 1 as CGFloat, green: 0.4, blue: 1, alpha: 1)
+            v1.backgroundColor = UIColor(red: 1, green: 0.4, blue: 1, alpha: 1)
             let v2 = UIView(frame:v1.bounds)
-            v2.backgroundColor = UIColor(red: 0.5 as CGFloat, green: 1, blue: 0, alpha: 1)
+            v2.backgroundColor = UIColor(red: 0.5, green: 1, blue: 0, alpha: 1)
             mainview.addSubview(v1)
             v1.addSubview(v2)
 
@@ -93,9 +93,9 @@ extension CGRect {
             
         case 7:
             let v1 = UIView(frame:CGRect(113, 111, 132, 194))
-            v1.backgroundColor = UIColor(red: 1 as CGFloat, green: 0.4, blue: 1, alpha: 1)
+            v1.backgroundColor = UIColor(red: 1, green: 0.4, blue: 1, alpha: 1)
             let v2 = UIView(frame:v1.bounds.insetBy(dx: 10, dy: 10))
-            v2.backgroundColor = UIColor(red: 0.5 as CGFloat, green: 1, blue: 0, alpha: 1)
+            v2.backgroundColor = UIColor(red: 0.5, green: 1, blue: 0, alpha: 1)
             mainview.addSubview(v1)
             v1.addSubview(v2)
