+++ bk2ch01p040sizeClasses/bk2ch01p040sizeClasses/MyView.swift
@@ -14,11 +14,11 @@ class MyView: UIView {
     var c2 : NSLayoutConstraint!
         
     func configure() {
-        self.backgroundColor = UIColor(red: 1 as CGFloat, green: 0.4, blue: 1, alpha: 1)
+        self.backgroundColor = UIColor(red: 1, green: 0.4, blue: 1, alpha: 1)
         let v2 = UIView()
-        v2.backgroundColor = UIColor(red: 0.5 as CGFloat, green: 1, blue: 0, alpha: 1)
+        v2.backgroundColor = UIColor(red: 0.5, green: 1, blue: 0, alpha: 1)
         let v3 = UIView()
-        v3.backgroundColor = UIColor(red: 1 as CGFloat, green: 0, blue: 0, alpha: 1)
+        v3.backgroundColor = UIColor(red: 1, green: 0, blue: 0, alpha: 1)
         
         v2.translatesAutoresizingMaskIntoConstraints = false
         v3.translatesAutoresizingMaskIntoConstraints = false
