+++ bk2ch01p041designableView/bk2ch01p040sizeClasses/MyView.swift
@@ -26,11 +26,11 @@ The representation is not perfect but it's pretty good.
     }
     
     func configure() {
-        self.backgroundColor = UIColor(red: 1 as CGFloat, green: 0.4, blue: 1, alpha: 1)
+        self.backgroundColor = UIColor(red: 1, green: 0.4, blue: 1, alpha: 1)
         let v2 = UIView()
-        v2.backgroundColor = UIColor(red: 0.5 as CGFloat, green: 1, blue: 0, alpha: 1)
+        v2.backgroundColor = UIColor(red: 0.5, green: 1, blue: 0, alpha: 1)
         let v3 = UIView()
-        v3.backgroundColor = UIColor(red: 1 as CGFloat, green: 0, blue: 0, alpha: 1)
+        v3.backgroundColor = UIColor(red: 1, green: 0, blue: 0, alpha: 1)
         
         v2.translatesAutoresizingMaskIntoConstraints = false
         v3.translatesAutoresizingMaskIntoConstraints = false
@@ -52,7 +52,7 @@ The representation is not perfect but it's pretty good.
     override func prepareForInterfaceBuilder() {
         // IB-only preparations can go here
         // typically this will involve supplying stub data
-        self.backgroundColor = UIColor(red: 1 as CGFloat, green: 0.4, blue: 1, alpha: 1)
+        self.backgroundColor = UIColor(red: 1, green: 0.4, blue: 1, alpha: 1)
         let lab = UILabel()
         lab.text = self.name
         lab.sizeToFit()
